+++ hastexo/migrations/0010_add_user_foreign_key.py
@@ -1,9 +1,30 @@
 from django.conf import settings
-from django.db import migrations, models
+from django.db import migrations, models, OperationalError
 import django.db.models.deletion
 
 
 class Migration(migrations.Migration):
+
+    def check_stacks(apps, schema_editor):
+        """
+        Check if all stacks can be linked to a real user account.
+        If not, error out with a suggestion how to proceed.
+        """
+        Stack = apps.get_model("hastexo", "Stack")
+        AnonymousUserId = apps.get_model("student", "AnonymousUserId")
+        anonymous_user_ids = AnonymousUserId.objects.values(
+            'anonymous_user_id').distinct()
+        problematic_stacks = Stack.objects.exclude(
+            student_id__in=anonymous_user_ids)
+
+        if len(problematic_stacks) > 0:
+            raise OperationalError(
+                'Unable to link stacks to users; please make sure that '
+                'the following stacks have a student_id that corresponds '
+                'to a real user account: '
+                f'{[s.name for s in problematic_stacks]}. '
+                'Please update or delete the stacks manually and rerun '
+                'the migration.')
 
     def backfill_learner(apps, schema_editor):
         """
@@ -22,6 +43,7 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
+        migrations.RunPython(check_stacks),
         migrations.AddField(
             model_name='stack',
             name='learner',